% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacter.R
\name{as.character}
\alias{as.character}
\alias{as.character,FittedFunctionalModel-method}
\alias{as.character,TransformedFittedFunctionalModel-method}
\title{Convert a \code{\link{FittedFunctionalModel}} to a String}
\arguments{
\item{x}{the object}

\item{x}{the object}
}
\value{
the name of the object

the name of the object
}
\description{
the \code{as.character} implementation for
  \code{\link{FittedFunctionalModel}}

the \code{as.character} implementation for
  \code{\link{TransformedFittedFunctionalModel}}
}
