% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R
\docType{class}
\name{FittedFunctionalModel-class}
\alias{FittedFunctionalModel-class}
\alias{FittedFunctionalModel}
\title{A Fitted Functional Model}
\description{
This class holds a fully parameterized, fitted functional model.
Instances of this class are typically the result of a model fitting approach,
i.e., the application of an algorithm which accepts an instance of
\code{\link{FunctionalModel}} and an instance of
\code{\link{RegressionQualityMetric}} and finds the parameters for the
functional model which minimize the return value of the regression quality
metric.
}
\section{Slots}{

\describe{
\item{\code{model}}{the model Blueprint}

\item{\code{par}}{the model parameters}
}}

\seealso{
FittedFunctionalModel.new
}
