% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacter.R
\name{FittedFunctionalModel.as.character}
\alias{FittedFunctionalModel.as.character}
\title{Convert a \code{FittedFunctionalModel} to a String}
\usage{
FittedFunctionalModel.as.character(x)
}
\arguments{
\item{x}{the \code{FittedFunctionalModel}}
}
\value{
the string
}
\description{
well, convert a \code{FittedFunctionalModel} to a String
}
