% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R
\name{FittedFunctionalModel.finalize}
\alias{FittedFunctionalModel.finalize}
\title{Finalize a Fitted Functional Model}
\usage{
FittedFunctionalModel.finalize(object)
}
\arguments{
\item{object}{the fitted model to be finalized}
}
\value{
a \code{\link{FittedFunctionalModel}}
}
\description{
This method makes sure that the created instance of
  \code{\link{FittedFunctionalModel}} is fully finalized. This involves
  substituting the function parameter into the function.
}
