% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R
\name{FittedFunctionalModel.new}
\alias{FittedFunctionalModel.new}
\title{Helper Method to Create a Fitted Functional Model}
\usage{
FittedFunctionalModel.new(model, par, quality)
}
\arguments{
\item{model}{the model \code{\link{FunctionalModel}}}

\item{par}{the model parameters}

\item{quality}{the quality of the model on the original data, computed by a
quality metric, smaller values are better}
}
\value{
a \code{\link{FittedFunctionalModel}}
}
\description{
Always use this function to instantiate
  \code{\link{FittedFunctionalModel}}
}
\seealso{
FittedFunctionalModel.finalized
}
