% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLearners.R
\name{FunctionalModel.defaultLearners}
\alias{FunctionalModel.defaultLearners}
\title{Get the Default Learners}
\usage{
FunctionalModel.defaultLearners()
}
\value{
a list of default learners that can be applied in
  \code{\link[regressoR.base]{regressoR.applyLearners}}.
}
\description{
Get the default learners for the models generated by
  \code{\link[regressoR.functional.models]{FunctionalModel.default}} for use
  in \code{\link[regressoR.base]{regressoR.applyLearners}}.
}
\seealso{
FunctionalModel.makeLearners
}
