% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultFitters.R
\name{FunctionalModel.fit.defaultFitters}
\alias{FunctionalModel.fit.defaultFitters}
\title{Get the List of Default Functional Model Fitting Algorithms}
\usage{
FunctionalModel.fit.defaultFitters(dataSize = 1000, paramCount = 4)
}
\arguments{
\item{dataSize}{the size of the data, assumed to be {@code 1000} if unknown}

\item{paramCount}{the number of model parameters, assumed to be {@code 4} if
unknown}
}
\value{
the List of Default Functional Model Fitters
}
\description{
Well, get the List of Default Functional Model Fitters.
}
