% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfoptim.R
\name{FunctionalModel.fit.dfoptim}
\alias{FunctionalModel.fit.dfoptim}
\title{Use Derivative-Free Local Searches to Optimize the Parameters}
\usage{
FunctionalModel.fit.dfoptim(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of
\code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of \code{\link{FittedFunctionalModel}}.
  \code{NULL} on failure.
}
\description{
Apply twi derivative-free algorithms, namely Hook-Jeeves and
  Nelder-Mead, to fit a functional model.
}
