% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{FunctionalModel.fit.lbfgsb}
\alias{FunctionalModel.fit.lbfgsb}
\title{Use the L-BFGS-B Algorithm from the \code{\link{optim}} Method of the
  \code{stats} Package for Fitting a Model}
\usage{
FunctionalModel.fit.lbfgsb(metric, model, par = NULL, q = 0.75)
}
\arguments{
\item{metric}{an instance of
\code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}

\item{q}{the effort to spent in learning, a value between 0 (min) and 1
(max). Higher values may lead to much more computational time, lower values
to potentially lower result quality.}
}
\value{
On success, an instance of \code{\link{FittedFunctionalModel}}.
  \code{NULL} on failure.
}
\description{
Apply the "L-BFGS-B" algorithm provided by the
  \code{\link{optim}} method from the \code{stats} package. Although the
  \code{\link{optim}} method should support using gradients, I just did not
  find out how to get this to work.
}
