% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls.R
\name{FunctionalModel.fit.nls}
\alias{FunctionalModel.fit.nls}
\title{Apply a the Standard Non-Linear Least Squares Approach from Package
  \code{stats}}
\usage{
FunctionalModel.fit.nls(metric, model, par = NULL, q = 0.75)
}
\arguments{
\item{metric}{an instance of
\code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}

\item{q}{the effort to spent in learning, a value between 0 (min) and 1
(max). Higher values may lead to much more computational time, lower values
to potentially lower result quality.}
}
\value{
On success, an instance of \code{\link{FittedFunctionalModel}}.
  \code{NULL} on failure.
}
\description{
Apply the default non-linear least squares algorithm to fit a
  functional model.
}
\seealso{
\code{\link{nls}}
}
