% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{FunctionalModel.fit.pso}
\alias{FunctionalModel.fit.pso}
\title{Use the \code{\link{psoptim}} Method from the \code{pso} Package for
  Fitting a Model}
\usage{
FunctionalModel.fit.pso(metric, model, par = NULL, q = 0.75)
}
\arguments{
\item{metric}{an instance of
\code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}

\item{q}{the effort to spent in learning, a value between 0 (min) and 1
(max). Higher values may lead to much more computational time, lower values
to potentially lower result quality.}
}
\value{
On success, an instance of \code{\link{FittedFunctionalModel}}.
  \code{NULL} on failure.
}
\description{
Apply the Particle Swarm Optimization algorithm to fit the
  parameters of a model.
}
