% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLearners.R
\name{FunctionalModel.makeLearners}
\alias{FunctionalModel.makeLearners}
\title{Create Learners for the Given Set of Models}
\usage{
FunctionalModel.makeLearners(models = FunctionalModel.all(),
  fitters = .defFit)
}
\arguments{
\item{models}{the set of models to use, by default this is
\code{\link{FunctionalModel.all}}}

\item{fitters}{the algorithm selecting which fitters to apply for each
model and data set, by default this is
\code{\link{FunctionalModel.fit.defaultFitters}}.}
}
\value{
a list of fitters that can be applied.
}
\description{
Create the learners for the use in
  \code{\link{regressoR.applyLearners}} from a set of models.
}
