% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLearners.R
\name{FunctionalModel.monotonousLearners}
\alias{FunctionalModel.monotonousLearners}
\title{Get the Default Learners for Monotonously Decreasing Models}
\usage{
FunctionalModel.monotonousLearners()
}
\value{
a list of default learners for models representing monotonously
  decreasing functions that can be applied in
  \code{\link[regressoR.base]{regressoR.applyLearners}}.
}
\description{
Get the default learners for the models generated by
  \code{\link[regressoR.functional.models]{FunctionalModel.monotonous}} for
  use in \code{\link[regressoR.base]{regressoR.applyLearners}}.
}
\seealso{
FunctionalModel.makeLearners
}
