% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R,
%   R/TransformedFittedFunctionalModel.R
\docType{methods}
\name{learning.Result.finalize}
\alias{learning.Result.finalize}
\alias{learning.Result.finalize,FittedFunctionalModel-method}
\alias{learning.Result.finalize,TransformedFittedFunctionalModel-method}
\title{Finalize a \code{\link{FittedFunctionalModel}}}
\usage{
\S4method{learning.Result.finalize}{FittedFunctionalModel}(object)

\S4method{learning.Result.finalize}{TransformedFittedFunctionalModel}(object)
}
\description{
Finalize a \code{\link{FittedFunctionalModel}}

Finalize a \code{\link{TransformedFittedFunctionalModel}}
}
