% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaes.R
\name{model.fit.cmaes}
\alias{model.fit.cmaes}
\title{Use CMA-ES to Optimize the Parameters}
\usage{
model.fit.cmaes(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
Apply the CMA-ES algorithm to fit a functional model.
}
