% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{model.fit.de}
\alias{model.fit.de}
\title{Use Differential Evolution to Optimize the Parameters}
\usage{
model.fit.de(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
Apply the Differential Evolution (DE) algorithm to fit a
  functional model.
}
