% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minqa.R
\name{model.fit.minqa}
\alias{model.fit.minqa}
\title{Use Powell's BOBYQA Approach to Optimize the Parameters}
\usage{
model.fit.minqa(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
Apply Powell's BOBYQA Approach to fit a functional model.
}
