% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls.R
\name{model.fit.nls}
\alias{model.fit.nls}
\title{Apply a the Standard Non-Linear Least Squares Approach from Package \code{stats}}
\usage{
model.fit.nls(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
Apply the default non-linear least squares algorithm to fit a
  functional model.
}
\seealso{
\code{\link{nls}}
}
