% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlslm.R
\name{model.fit.nlslm}
\alias{model.fit.nlslm}
\title{Apply a Levenberg-Marquardt Algorithm}
\usage{
model.fit.nlslm(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{regressoR.quality::RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
Apply the Levenberg-Marquardt algorithm to fit a functional
  model.
}
