% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacter.R
\name{TransformedFittedFunctionalModel.as.character}
\alias{TransformedFittedFunctionalModel.as.character}
\title{Convert a \code{TransformedFittedFunctionalModel} to a String}
\usage{
TransformedFittedFunctionalModel.as.character(x)
}
\arguments{
\item{x}{the \code{TransformedFittedFunctionalModel}}
}
\value{
the string
}
\description{
well, convert a \code{TransformedFittedFunctionalModel} to a String
}
