% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformedFittedFunctionalModel.R
\name{TransformedFittedFunctionalModel.finalize}
\alias{TransformedFittedFunctionalModel.finalize}
\title{Finalize a Fitted Transformed Functional Model}
\usage{
TransformedFittedFunctionalModel.finalize(object)
}
\arguments{
\item{object}{the fitted model to be finalized}
}
\value{
a \code{\link{TransformedFittedFunctionalModel}} or
  \code{\link{FittedFunctionalModel}}
}
\description{
This method makes sure that the created instance of
  \code{\link{TransformedFittedFunctionalModel}} is fully finalized. This
  involves substituting the function parameter into the function.
}
