% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R
\docType{class}
\name{FittedFunctionalModel-class}
\alias{FittedFunctionalModel-class}
\alias{FittedFunctionalModel}
\title{A Fitted Functional Model}
\description{
This class holds a fully parameterized, fitted functional model.
}
\section{Slots}{

\describe{
\item{\code{model}}{the model Blueprint}

\item{\code{par}}{the model parameters}
}}

\seealso{
FittedFunctionalModel.new
}
