% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{FunctionalModel.fit}
\alias{FunctionalModel.fit}
\title{Fit the Given Model Blueprint to the Specified Data}
\usage{
FunctionalModel.fit(metric, model, par = NULL)
}
\arguments{
\item{metric}{an instance of \code{RegressionQualityMetric}}

\item{model}{an instance of \code{\link{FunctionalModel}}}

\item{par}{the initial starting point}
}
\value{
On success, an instance of
  \code{\link{FittedFunctionalModel}}. \code{NULL} on failure.
}
\description{
This is the general method to be used to fit a \code{\link{FunctionalModel}} to
a \code{\link{RegressionQualityMetric}}.
}
