% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedFunctionalModel.R,
%   R/TransformedFittedFunctionalModel.R
\name{learning.Result.finalize}
\alias{learning.Result.finalize}
\alias{learning.Result.finalize,FittedFunctionalModel-method}
\alias{learning.Result.finalize,TransformedFittedFunctionalModel-method}
\title{Finalize a Fitted Functional Model}
\arguments{
\item{object}{the fitted model to be finalized}

\item{object}{the fitted model to be finalized}
}
\value{
a \code{\link{FittedFunctionalModel}}

a \code{\link{TransformedFittedFunctionalModel}}
}
\description{
This method makes sure that the created instance of
  \code{\link{FittedFunctionalModel}} is fully finalized. This involves
  substituting the function parameter into the function.

This method makes sure that the created instance of
  \code{\link{TransformedFittedFunctionalModel}} is fully finalized. This
  involves substituting the function parameter into the function.
}
